Add square collision detection
Added detect_collision which checks the square area around the marble and compares those positions to map. Right now it only output to the cmd window if it detects a collision. It also alerts if the bounds of the marble fall outside the 600x600 plane. Still need to change from a square area to check to the actual circle

Changed update_tick so it enforces integers only for position.

Added a feature that kills the script if the marble falls outside of the 600x600 plane

Changed radius for draw function to 5

Changed accelerations in main for testing

+++ update_tick.m
@@ -3,11 +3,11 @@
 function [position, linear_velocity, linear_acceleration, angular_velocity, angular_acceleration] = update_tick(position, linear_velocity, linear_acceleration, angular_velocity, angular_acceleration)
 
     if linear_velocity(2) ~= 0 %%if y linear velocity is not zero then add that velocity to y position
-        position(2) = position(2) + linear_velocity(2)/10; %%divide by 10 because of our time interval
+        position(2) = int16(position(2) + linear_velocity(2)/10); %%divide by 10 because of our time interval
     end
     
     if linear_velocity(1) ~= 0 %%if x linear velocity is not zero then add that velocity to x position
-        position(1) = position(1) + linear_velocity(1)/10; %%divide by 10 because of our time interval
+        position(1) = int16(position(1) + linear_velocity(1)/10); %%divide by 10 because of our time interval
     end
 
     if angular_acceleration(2) ~= 0 %%if y angular acceleration is not zero then add that acceleration to y angular velocity
